rename rules -> pipeline
