+++ src/functions/pipeline/createRule.js
