+++ src/functions/pipeline/deleteById.js
