+++ src/functions/pipeline/removeEnv.js
