+++ src/functions/pipeline/ruleById.js
