+++ src/functions/pipeline/templates.js
