+++ src/functions/pipeline/updateRule.js
